highlight_node highlight the nodes

highlight_node highlight the nodes

Usage

highlight_node(
  node_name = NULL,
  highlight_color = NULL,
  filter = NULL,
  glow = FALSE,
  glow_base_size = FALSE,
  shape_number = NULL,
  shape_size = NULL,
  shape_color = NULL,
  shape_with_text = NULL,
  with_text = NULL,
  text_attribute = NULL,
  text_node_color = FALSE,
  glow_size = 1.2,
  override_text = FALSE,
  use_ggfx = NULL,
  ggfx_params = list(),
  textpath_params = list(),
  circ_size = 1.1
)

Arguments

node_name

used to subset node data with `name` argument

highlight_color

highlighting color

filter

used to subset node data specified as string

glow

use glow highlighting

glow_base_size

use base size of the node for initial size of glowing

shape_number

shape number

shape_size

shape size

shape_color

shape color

shape_with_text

if specified, show text along with shape (21 only) using `geomtextpath`

with_text

if specified, show text around the nodes, not the shape

text_attribute

if not NULL, graph attributes of this name will be used as text annotation

text_node_color

if TRUE, node color will be used as text color

glow_size

argument to control how big the glowing will be

override_text

after highlighting, stack the last geom_node_text layer

use_ggfx

use ggfx geom to highlight, default to NULL

ggfx_params

ggfx parameters

circ_size

circle size for text highlighting